Moved common logic to different class
